+++ CHANGELOG.md
@@ -24,8 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     bits respectively.
     
 - Added new file `pattern_algorithms.py`:
-  – Introduces the `compute_common_fixedmask` function, which computes the bitwise AND (intersection) of the fixedmask values from a list of Pattern objects.
-  - Added the `generate_tree_by_common_bits` method. This function generates a hierarchical tree structure of Patterns by computing a common fixed bit mask and recursively grouping patterns based on their fixed bits. Each group is represented as either a leaf or a tree node depending on the significance of the common fixed bits.
+  – Added the `compute_common_fixedmask` function, which computes the bitwise AND (intersection) of the fixedmask values from a list of Pattern objects.
+  - Added the `compute_fixed_bit_groups` function, which groups patterns by their fixed part (the key) and associates each group with tuples of (remaining part, original pattern).
+  - Added the `generate_tree_by_common_bits` function, which constructs a hierarchical tree of Patterns. It computes a common fixed bit mask, then recursively groups 
+    Patterns based on their matching fixed bits, representing each group as either a leaf or a tree node depending on the significance of the fixed bits.
 
 - Added the `Printer` class in `printer.py` and the `IPrinter` interface in `i_printer.py`. Both are designed to abstract the print command for more testable use cases.